OncRpcException: Remove message override. Add constructor with additional information message. Set message type as Enum.  Add constructor with inner exception.

+++ src/rpc/oncrpc/ONC.RPC/Client/OncRpcClientCallMessage.cs
@@ -33,7 +33,7 @@ public class OncRpcClientCallMessage : OncRpcCallMessageBase
     public virtual void Encode( XdrEncodingStreamBase encoder )
     {
         encoder.EncodeInt( this.MessageId );
-        encoder.EncodeInt( this.MessageType );
+        encoder.EncodeInt( ( int ) this.MessageType );
         encoder.EncodeInt( this.ProtocolVersion );
         encoder.EncodeInt( this.Program );
         encoder.EncodeInt( this.Version );
